requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiofiles==23.2.1
    # via aiogram
aiogram==3.5.0
    # via -r requirements.in
aiohttp==3.9.5
    # via aiogram
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
attrs==23.2.0
    # via aiohttp
certifi==2024.2.2
    # via aiogram
dnspython==2.6.1
    # via pymongo
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
idna==3.7
    # via yarl
magic-filter==1.0.12
    # via aiogram
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
pydantic==2.7.1
    # via aiogram
pydantic-core==2.18.2
    # via pydantic
pymongo==4.7.1
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via -r requirements.in
typing-extensions==4.11.0
    # via
    #   aiogram
    #   pydantic
    #   pydantic-core
yarl==1.9.4
    # via aiohttp
